some editorial visual changes
most .html flourishes do not work in .md files, only in .rmd that are rendered. Also, I removed a bunch of blank space from the table - this is mostly a suggestion. I know that this is how it comes out of R Studio, it is just kind of impossible to read in an .md doc outside of R Studio.